+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.0.35] - 2021-03-15
+
 ### Changed
 
 - Changed Marketplace setting to dropdown